refactor(api): centralize safe upstream response parsing
