+++ hushh-webapp/app/api/consent/export-refresh/fail/route.ts
@@ -1,24 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
-
+import { safeParseResponse } from "@/app/api/_utils/safe-response";
 import { getPythonApiUrl } from "@/app/api/_utils/backend";
 
 const BACKEND_URL = getPythonApiUrl();
 
-async function safeParseResponse(response: Response) {
-  const contentType = response.headers.get("content-type") || "";
 
-  if (!contentType.includes("application/json")) {
-    const text = await response.text().catch(() => "");
-    return text ? { error: text } : {};
-  }
-
-  try {
-    return await response.json();
-  } catch {
-    const text = await response.text().catch(() => "");
-    return text ? { error: text } : {};
-  }
-}
 
 export async function POST(request: NextRequest) {
   try {
