+++ hushh-webapp/app/api/consent/export-refresh/jobs/route.ts
@@ -1,17 +1,24 @@
 import { NextRequest, NextResponse } from "next/server";
-
+import { safeParseResponse } from "@/app/api/_utils/safe-response";
 import { getPythonApiUrl } from "@/app/api/_utils/backend";
 
 const BACKEND_URL = getPythonApiUrl();
 
+
+
 export async function GET(request: NextRequest) {
   try {
     const userId = request.nextUrl.searchParams.get("userId") || "";
+
     if (!userId) {
-      return NextResponse.json({ error: "userId is required" }, { status: 400 });
+      return NextResponse.json(
+        { error: "userId is required" },
+        { status: 400 }
+      );
     }
 
     const authHeader = request.headers.get("Authorization");
+
     if (!authHeader) {
       return NextResponse.json(
         { error: "Authorization header required" },
@@ -27,12 +34,18 @@ export async function GET(request: NextRequest) {
         },
       }
     );
-    const payload = await response
-      .json()
-      .catch(async () => ({ error: await response.text().catch(() => "") }));
-    return NextResponse.json(payload, { status: response.status });
+
+    const payload = await safeParseResponse(response);
+
+    return NextResponse.json(payload, {
+      status: response.status,
+    });
   } catch (error) {
     console.error("[API] export-refresh/jobs error:", error);
-    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
+
+    return NextResponse.json(
+      { error: "Internal server error" },
+      { status: 500 }
+    );
   }
 }
